Release: Version 1.30
There have not been a ipvsadm release for far too long. This release is
focused on kernel release v4.18 and below, for adding the missing userspace
side for the schedulers MH, FO and OVF. There are upcoming change for GUE
and GRE, that are not part of this release.

There have been very little development on the tool, but this release still
contains userspace config to kernel side features that span many kernel
releases. Special thanks to Quentin Armitage for adding this userspace side
support, that was missing in ipvsadm.

This release contains userspace support and/or doc for 3 more schedulers:
 - mh:  Maglev Hashing: added in kernel v4.18
 - ovf (doc): Weighted Overflow: added in kernel v4.3
   - https://git.kernel.org/torvalds/c/eefa32d3f3c5
 - fo (doc): Weighted FailOver: added in kernel v3.18
   - https://git.kernel.org/torvalds/c/616a9be25cb9

This release is based on the kernel.org git tree:
  https://git.kernel.org/cgit/utils/kernel/ipvsadm/ipvsadm.git/

You can download the tarballs from:
 https://kernel.org/pub/linux/utils/kernel/ipvsadm/

Git tree:
 git://git.kernel.org/pub/scm/utils/kernel/ipvsadm/ipvsadm.git

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>